slash gateway flow complete

+++ src/constant.rs
@@ -1,2 +1,2 @@
 pub const REQUEST_RELAY_TIMEOUT: u64 = 15 * 60;
-pub const MAX_GATEWAY_RETRIES: u64 = 2;
+pub const MAX_GATEWAY_RETRIES: u8 = 2;
